# Conflicts:
#   src/services/contract-generator.service.ts

+++ src/services/contract-generator.service.ts
@@ -31,7 +31,7 @@ class ContractGenerator {
     return tempDir;
   }
 
-  async generateContract(collectionSettings : Collection, enums?: EnumParameter[], mediafiles?: MediaFile[]){
+  async generateContract(collectionSettings : Collection, jsonCollection: string, enums?: EnumParameter[], mediafiles?: MediaFile[]){
     const hashContract = sha256(JSON.stringify(collectionSettings));
     const tempDir = path.join(globals.RESULT_COLLECTION, hashContract);
 
@@ -53,34 +53,39 @@ class ContractGenerator {
     const enumsFileTemp = path.join(tempDir, 'libraries', 'Enums.sol');
     const debotFileTemp = path.join(tempDir, 'debots', 'MintingDebot.sol');
 
-    fs.mkdirSync(tempDir);
+    if (!fs.existsSync(tempDir)) {
+      fs.mkdirSync(tempDir);
+      fs.cpSync(interfacesDir, interfacesDirTepm, {recursive: true});
+      fs.cpSync(librariesDir, librariesDirTepm, {recursive: true});
+      fs.cpSync(resolversDir, resolversDirTepm, {recursive: true});
+      fs.cpSync(debotLibDir, debotLibDirTepm, {recursive: true});
+      fs.cpSync(debotsDir, debotsDirTepm, {recursive: true});
+      fs.copyFileSync(indexFile, indexFileTepm);
+      fs.copyFileSync(indexBasisFile, indexBasisFileTepm);
+      fs.copyFileSync(directSaleRootFile, directSaleRootFileTemp);
+      fs.copyFileSync(directSaleFile, directSaleFileTemp);
+      fs.copyFileSync(auctionRootFile, auctionRootFileTemp);
+      fs.copyFileSync(auctionFile, auctionFileTemp);
+      fs.copyFileSync(aDataCoreFile, aDataCoreFileTemp);
+      fs.writeFileSync(path.join(tempDir, "collectionInfo.json"), jsonCollection, {flag: 'w'});
 
-    fs.cpSync(interfacesDir, interfacesDirTepm, {recursive: true});
-    fs.cpSync(librariesDir, librariesDirTepm, {recursive: true});
-    fs.cpSync(resolversDir, resolversDirTepm, {recursive: true});
-    fs.cpSync(debotLibDir, debotLibDirTepm, {recursive: true});
-    fs.cpSync(debotsDir, debotsDirTepm, {recursive: true});
-    fs.copyFileSync(indexFile, indexFileTepm);
-    fs.copyFileSync(indexBasisFile, indexBasisFileTepm);
-    fs.copyFileSync(directSaleRootFile, directSaleRootFileTemp);
-    fs.copyFileSync(directSaleFile, directSaleFileTemp);
-    fs.copyFileSync(auctionRootFile, auctionRootFileTemp);
-    fs.copyFileSync(auctionFile, auctionFileTemp);
-    fs.copyFileSync(aDataCoreFile, aDataCoreFileTemp);
+      let addParamsService = new AddParamsService();
 
-    let addParamsService = new AddParamsService();
-
-    if (collectionSettings.getParameters().length == 0) {
-      fs.copyFileSync(nftRootFile, nftRootFileTepm);
-      fs.copyFileSync(dataFile, dataFileTepm);
-    } else {
-      await addParamsService.addSeveralParams(collectionSettings.getParameters(), nftRootFile, nftRootFileTepm);
-      await addParamsService.addSeveralParams(collectionSettings.getParameters(), dataFile, dataFileTepm);
-      await addParamsService.addSeveralParams(collectionSettings.getParameters(), iDataFileTemp, iDataFileTemp);
-      await addParamsService.addSeveralParams(collectionSettings.getParameters(), debotFileTemp, debotFileTemp);
-      if (enums !== undefined) {
-        await addParamsService.addEnums(enums, enumsFileTemp, enumsFileTemp);
-        await addParamsService.addEnums(enums, debotFileTemp, debotFileTemp);
+      if (collectionSettings.getParameters().length == 0) {
+        fs.copyFileSync(nftRootFile, nftRootFileTepm);
+        fs.copyFileSync(dataFile, dataFileTepm);
+      } else {
+        await addParamsService.addSeveralParams(collectionSettings.getParameters(), nftRootFile, nftRootFileTepm);
+        await addParamsService.addSeveralParams(collectionSettings.getParameters(), dataFile, dataFileTepm);
+        await addParamsService.addSeveralParams(collectionSettings.getParameters(), iDataFileTemp, iDataFileTemp);
+        await addParamsService.addSeveralParams(collectionSettings.getParameters(), debotFileTemp, debotFileTemp);
+        if (enums !== undefined) {
+          await addParamsService.addEnums(enums, enumsFileTemp, enumsFileTemp);
+          await addParamsService.addEnums(enums, debotFileTemp, debotFileTemp);
+        }
+        if (mediafiles !== undefined) {
+          await addParamsService.addMediaFiles(mediafiles, debotFileTemp, debotFileTemp);
+        }
       }
       if (mediafiles !== undefined) {
         await addParamsService.addMediaFiles(mediafiles, debotFileTemp, debotFileTemp);
